added UserService, UnauthorizedAccessException class, associate User with Transaction

+++ src/main/java/pl/madej/finansemanangerrestapi/service/TransactionService.java
@@ -1,17 +1,21 @@
 package pl.madej.finansemanangerrestapi.service;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import pl.madej.finansemanangerrestapi.error.TransactionNotFoundException;
+import pl.madej.finansemanangerrestapi.error.UnauthorizedAccessException;
 import pl.madej.finansemanangerrestapi.mapper.TransactionMapper;
 import pl.madej.finansemanangerrestapi.model.Transaction;
 import pl.madej.finansemanangerrestapi.model.User;
 import pl.madej.finansemanangerrestapi.payload.transaction.TransactionRequest;
 import pl.madej.finansemanangerrestapi.payload.transaction.TransactionResponse;
 import pl.madej.finansemanangerrestapi.repository.TransactionRepository;
+import pl.madej.finansemanangerrestapi.repository.UserRepository;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -19,20 +23,34 @@ import java.util.stream.Collectors;
 public class TransactionService {
 
     private final TransactionRepository transactionRepository;
+    private final UserService userService;
 
     public Long addTransaction(TransactionRequest transactionRequest) {
 
         Transaction transaction = TransactionMapper.INSTANCE.toTransaction(transactionRequest);
         transaction.setDate(LocalDateTime.now());
-        transaction.setUser(new User());
+
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+
+        User loggedUser = userService.getUserByUsername(username);
+
+        transaction.setUser(loggedUser);
 
         return transactionRepository.save(transaction).getId();
     }
 
     public void deleteTransaction(Long transactionId) {
-        transactionRepository.findById(transactionId)
+        Transaction transaction = transactionRepository.findById(transactionId)
                 .orElseThrow(() -> new TransactionNotFoundException("Transaction not found with id " + transactionId));
-        transactionRepository.deleteById(transactionId);
+
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        User loggedUser = userService.getUserByUsername(username);
+
+        if (!transaction.getUser().equals(loggedUser)) {
+            throw new UnauthorizedAccessException("User does not have permission to delete this transaction.");
+        }
+
+            transactionRepository.deleteById(transactionId);
     }
 
     public TransactionResponse updateTransaction(Long transactionId, TransactionRequest transactionRequest) {
